more pr feedback
