+++ .gitignore
@@ -19,6 +19,7 @@ xcuserdata
 DerivedData
 .idea/
 *.LSOverride
+xcshareddata
 
 # Cocoapods
